+++ lib/components/DateTime/DateTimeField.tsx
@@ -78,6 +78,9 @@ export interface DateTimeFieldProps extends React.Props<DateTimeFieldType> {
      */
     onExpand?: (expanded: boolean) => void;
 
+    /** React node to render at the far side of the label. */
+    labelFarSide?: React.ReactNode;
+
     attr?: DateTimeFieldAttributes & FormFieldAttributes;
 }
 
@@ -357,6 +360,7 @@ export class DateTimeField extends React.Component<DateTimeFieldProps, Partial<D
                 className={css('datetime-field', this.props.className)}
                 attr={fieldAttr}
                 tooltip={this.props.tooltip}
+                labelFarSide={this.props.labelFarSide}
             >
                 <Attr.div
                     className={css('field-content')}