+++ lib/components/Field/FormLabel.tsx
@@ -10,6 +10,7 @@ export interface FormLabelType {}
 
 export interface FormLabelAttributes {
     container?: DivProps;
+    innerContainer?: DivProps;
     text?: LabelProps;
     icon?: IconAttributes;
     balloon?: BalloonAttributes;
@@ -58,6 +59,9 @@ export interface FormLabelProps extends React.Props<FormLabelType> {
     /** Classname to append to top level element */
     className?: string;
 
+    /** Extra node to render at the far side of the label */
+    farSide?: React.ReactNode;
+
     attr?: FormLabelAttributes;
 }
 
@@ -102,14 +106,23 @@ export const FormLabel: React.StatelessComponent<FormLabelProps> = (props: FormL
             className={css('label-container', props.className)}
             attr={props.attr.container}
         >
-            <Attr.label
-                className={css('label', {'required': props.required})}
-                htmlFor={props.name}
-                attr={props.attr.text}
-            >
-                {props.children}
-            </Attr.label>
-            {balloon}
+            <Attr.div
+                className={css('label-inner-container')}
+                attr={props.attr.innerContainer}>
+                <Attr.label
+                    className={css('label', { 'required': props.required })}
+                    htmlFor={props.name}
+                    attr={props.attr.text}
+                >
+                    {props.children}
+                </Attr.label>
+                {balloon}
+            </Attr.div>
+            <Attr.div
+                className={css('label-inner-container')}
+                attr={props.attr.innerContainer}>
+                {props.farSide}
+            </Attr.div>
         </Attr.div>
     );
 };